requirements.txt
streamlit
tensorflow
numpy
pyrebase4
python-dotenv
pillow